+++ pkg/middleware/basic_session_test.go
@@ -26,6 +26,7 @@ var _ = Describe("Basic Auth Session Suite", func() {
 
 		type basicAuthSessionLoaderTableInput struct {
 			authorizationHeader string
+			preferEmail         bool
 			sessionGroups       []string
 			existingSession     *sessionsapi.SessionState
 			expectedSession     *sessionsapi.SessionState
@@ -55,7 +56,7 @@ var _ = Describe("Basic Auth Session Suite", func() {
 				// Create the handler with a next handler that will capture the session
 				// from the scope
 				var gotSession *sessionsapi.SessionState
-				handler := NewBasicAuthSessionLoader(validator, in.sessionGroups)(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+				handler := NewBasicAuthSessionLoader(validator, in.sessionGroups, in.preferEmail)(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 					gotSession = middlewareapi.GetRequestScope(r).Session
 				}))
 				handler.ServeHTTP(rw, req)
@@ -117,6 +118,12 @@ var _ = Describe("Basic Auth Session Suite", func() {
 				sessionGroups:       []string{"a", "b"},
 				existingSession:     nil,
 				expectedSession:     &sessionsapi.SessionState{User: "admin", Groups: []string{"a", "b"}},
+			}),
+			Entry("Basic Base64(user1:<user1Password>) (with PreferEmailToUser)", basicAuthSessionLoaderTableInput{
+				authorizationHeader: "Basic dXNlcjE6VXNFck9uM1A0NTU=",
+				preferEmail:         true,
+				existingSession:     nil,
+				expectedSession:     &sessionsapi.SessionState{User: "user1", Email: "user1"},
 			}),
 		)
 	})